Revert "Add repos overview page and navigation button"
