+++ src/plugins/router/routes.js
@@ -26,15 +26,4 @@ export const routes = [
       },
     ],
   },
-  {
-    path: '/repos',
-    component: () => import('@/layouts/default.vue'),
-    children: [
-      {
-        path: '',
-        name: 'Repos',
-        component: () => import('@/pages/repos.vue'),
-      },
-    ],
-  },
 ];
