+++ src/views/dashboard/Title.vue
@@ -1,6 +1,66 @@
 <script setup>
-import NavbarActions from '@/layouts/components/NavbarActions.vue';
-import NavbarBranding from '@/layouts/components/NavbarBranding.vue';
+import { ref, onMounted, onUnmounted, computed } from 'vue';
+import { useRouter } from 'vue-router';
+import NavbarThemeSwitcher from '@/layouts/components/NavbarThemeSwitcher.vue';
+
+const user = ref(null);
+const router = useRouter();
+const API_BASE = (import.meta.env.VITE_API_BASE_URL || 'https://ossprey.ngrok.app').replace(/\/$/, '');
+
+// Reactive variable to track viewport width
+const isMobileView = ref(window.innerWidth < 768);
+
+// Function to update viewport state
+const updateViewport = () => {
+  isMobileView.value = window.innerWidth < 768;
+};
+
+// Add and remove event listener on component lifecycle
+onMounted(() => {
+  window.addEventListener('resize', updateViewport);
+  const stored = localStorage.getItem('user');
+  if (stored) {
+    try {
+      user.value = JSON.parse(stored);
+    }
+    catch {
+      // ignore parse errors
+    }
+  }
+});
+
+onUnmounted(() => {
+  window.removeEventListener('resize', updateViewport);
+});
+
+const userName = computed(() => user.value?.name || user.value?.email || '');
+
+const logout = async () => {
+  const email = user.value?.email;
+  try {
+    await fetch(`${API_BASE}/api/logout`, { method: 'POST' });
+  }
+  catch {
+    // ignore errors
+  }
+
+  if (email) {
+    fetch(`${API_BASE}/api/track_logout`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ user_email: email }),
+    }).catch(err => {
+      console.error('Failed to track logout:', err);
+    });
+  }
+
+  localStorage.removeItem('user');
+  window.dispatchEvent(new Event('user-auth-changed'));
+  user.value = null;
+  router.push('/');
+};
 </script>
 
 <template>
@@ -9,14 +69,37 @@ import NavbarBranding from '@/layouts/components/NavbarBranding.vue';
       
       <!-- Logo + Title -->
       <div class="title-container">
-        <VCardTitle class="d-flex align-center mb-0">
-          <NavbarBranding />
+        <VCardTitle class="text-primary font-weight-bold d-flex align-center mb-0">
+          <img
+            src="/ospex-logo.png"
+            alt="OSPEx Logo"
+            style="height: 32px; width: auto; margin-right: 10px;"
+          />
+
+          <span v-if="isMobileView">OSSPREY</span>
+          <span v-else>OSSPREY (Open Source Software PRojEct sustainabilitY tracker)</span>
         </VCardTitle>
       </div>
 
       <!-- Action Icons -->
       <div class="actions-container">
-        <NavbarActions />
+        <NavbarThemeSwitcher />
+
+        <VTooltip v-if="user" :text="userName" location="bottom">
+          <template #activator="{ props }">
+            <VBtn v-bind="props" icon class="ms-2">
+              <VIcon icon="bx-user" />
+            </VBtn>
+          </template>
+        </VTooltip>
+
+        <VTooltip v-if="user" text="Log Out" location="bottom">
+          <template #activator="{ props }">
+            <VBtn v-bind="props" icon class="ms-2" @click="logout">
+              <VIcon icon="bx-log-out" />
+            </VBtn>
+          </template>
+        </VTooltip>
       </div>
     </VCardText>
   </VCard>
